[abstractHandCtrl] fix some MSVC++ warnings
svn path=/trunk/iCub/; revision=2036

+++ src/abstractHandCtrl/lib/src/Hand.cpp
@@ -446,7 +446,7 @@ void Hand::setSamplingRate(int t) {
 }
 
 int Hand::getSamplingRate() {
-	return recorder->getRate();
+	return (int) recorder->getRate();
 }
 
 void Hand::setMonitorRate(int i) {
